This is a big update..
First, I fixed some errors in the files, as usual, and then I worked more on the field centric teleop, and I used odometry wheels to solve for the robot heading. I installed a library called FtcLib to do this, and on the robot we will have to put 3 odometry pods, 2 parallel, and 1 in the center which is perpendicular. Since in the field centric teleop, we are getting measurements from both the IMU and the odometry for the robot heading, I started on making a filter to combine the 2 readings and estimate the heading. I need to understand the Filter code that is there so far....